requirements.txt
absl-py==0.6.1
anykeystore==0.2
apex==0.1
arxiv-latex-cleaner==0.1.10
astor==0.8.0
better-exceptions==0.2.2
bpytop==1.0.25
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
cryptacular==1.5.5
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
dirt==0.3.0
dominate==2.5.2
fsspec==0.8.3
future==0.18.2
gast==0.2.2
glumpy==1.1.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
googledrivedownloader==0.4
grpcio==1.32.0
h5py==2.10.0
hupper==1.10.2
idna==2.9
-e git+https://github.com/jackd/ifp-sample@d506204da70f2b33c4301b2378b0d623c2a4802a#egg=ifp_sample
imageio==2.6.1
isodate==0.6.0
joblib==0.13.2
jsonpatch==1.26
jsonpointer==2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Lasagne==0.2.dev1
llvmlite==0.31.0
lmdb==0.98
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mpmath==1.1.0
networkx==2.4
nose==1.3.7
numba==0.48.0
numpy==1.17.2
oauthlib==3.1.0
opencv-python==4.2.0.32
openmesh==1.1.3
opt-einsum==3.1.0
pandas==1.0.3
PasteDeploy==2.1.0
pbkdf2==1.3
Pillow==7.2.0
pkg-resources==0.0.0
plaster==1.0
plaster-pastedeploy==0.7
plyfile==0.7.2
point-cloud-utils==0.15.1
portalocker==1.7.0
protobuf==3.9.1
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycollada==0.7.1
Pygments==2.7.1
PyMCubes==0.0.12
pymesh2==0.2.1
pyobb==1.0.2
PyOpenGL==3.1.4
pyparsing==2.4.5
pyramid==1.10.4
pyramid-mailer==0.15.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytorch-lightning==1.0.1
pytorch-msssim==0.2.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
rdflib==4.2.2
repoze.sendmail==4.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.1.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.19
sympy==1.5.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Theano==1.0.5
torch==1.4.0
torch-cluster==1.5.3
torch-encoding==1.2.0b20200419
torch-geometric==1.4.3
torch-scatter==2.0.4
torch-sparse==0.6.1
torch-spline-conv==1.2.0
torchfile==0.1.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.43.0
transaction==3.0.0
translationstring==1.4
triangle==20190115.3
trimesh==3.2.17
urllib3==1.25.8
velruse==1.1.1
venusian==3.0.0
visdom==0.1.8.9
WebOb==1.8.6
websocket-client==0.57.0
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.3.3
wtforms-recaptcha==0.3.2
zlx==0.0.15
zope.deprecation==4.4.0
zope.interface==5.1.0
zope.sqlalchemy==1.3
